+++ ci/release_notes.md
@@ -1,1 +1,5 @@
-Empty - please add release notes here
+##Breaking changes
+- Change the schema of the persisted state file to differentiate between propagated and latest files
+
+##Bug Fix
+- Fix wording for `cepler concourse ci_out` command